Registration and BalanceCheck

+++ src/main/java/cap/hackyeah/coin_keeper_hackyeah_2024_back/dto/Player.java
@@ -29,7 +29,7 @@ public class Player {
 
     private String password;
 
-    private Integer balance;
+    private Long balance;
 
-    private Integer goal;
+    private Long goal;
 }
